Выделен PointController
